# Optimize some string handling