I've added an additional commit to this PR that further optimizes this area of code:

```none
benchmark                       old ns/op     new ns/op     delta
BenchmarkTreeToTomlString-2     49028         18817         -61.62%

benchmark                       old allocs     new allocs     delta
BenchmarkTreeToTomlString-2     485            78             -83.92%

benchmark                       old bytes     new bytes     delta
BenchmarkTreeToTomlString-2     9749          4961          -49.11%
```