@kevinburke 2017-05-31
Here's what I had in mind. It makes the code a lot simpler, since we know that `b.WriteString` and company never return an error, we don't need to check them constantly.

https://github.com/pelletier/go-toml/commit/b49b944a707ec81a11ce2f71249f8cf2bc32b5e4