@pelletier 2017-06-02
@moorereason nice!
@kevinburke I like your proposition (makes the code much cleaner), but the +14% allocs increase isn't great. What happens with a larger preallocated buffer?

I think I'm going to merge this as a first step. Then a second step we should figure a way to make Kevin's solution use some less allocs, as it makes the code much more readable.