@pelletier 2017-06-02
Ref https://github.com/pelletier/go-toml/issues/167